requirements.txt
googleapis_common_protos==1.52.0
pytube3==9.6.4
requests==2.24.0
Google==3.0.0
google_api_python_client==1.12.2
google_auth_oauthlib==0.4.1
gspread==3.6.0
more_itertools==8.5.0
PyYAML==5.3.1
questionary==1.5.2
simple_youtube_api==0.2.8
unicode_slugify==0.1.3
youtube_dl==2020.9.20
